requirements.txt
openai
requests
beautifulsoup4
numpy
scikit-learn
matplotlib
python-dotenv
